requirements.txt
jupyter
numpy
tqdm
pandas
seaborn
scikit-learn
matplotlib
opencv-python
pillow
torch
torchvision
streamlit
albumentations
timm
omegaconf
ultralytics
wandb
iglovikov_helper_functions
people_segmentation
lion_pytorch